# swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "API documentation for the store",
    "version": "1.0.0",
    "title": "Store API"
  },
  "host": "localhost:3000",
  "tags": [
    {
      "name": "orders",
      "description": "Order management"
    },
    {
      "name": "posts",
      "description": "News posts management"
    },
    {
      "name": "products",
      "description": "Product management"
    },
    {
      "name": "users",
      "description": "User management"
    }
  ],
  "paths": {
    "/orders": {
      "get": {
        "tags": ["orders"],
        "summary": "Get all orders",
        "responses": {
          "200": {
            "description": "A list of orders",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Order"
              }
            }
          }
        }
      },
      "post": {
        "tags": ["orders"],
        "summary": "Create a new order",
        "parameters": [
          {
            "name": "order",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Order"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Order created successfully",
            "schema": {
              "$ref": "#/definitions/Order"
            }
          }
        }
      },
      "put": {
        "tags": ["orders"],
        "summary": "Update an existing order",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "The ID of the order to update"
          },
          {
            "name": "order",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Order"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Order updated successfully",
            "schema": {
              "$ref": "#/definitions/Order"
            }
          }
        }
      },
      "delete": {
        "tags": ["orders"],
        "summary": "Delete an order",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "The ID of the order to delete"
          }
        ],
        "responses": {
          "200": {
            "description": "Order deleted successfully"
          }
        }
      }
    },
    "/posts": {
      "get": {
        "tags": ["posts"],
        "summary": "Get all posts",
        "responses": {
          "200": {
            "description": "A list of posts",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Post"
              }
            }
          }
        }
      },
      "post": {
        "tags": ["posts"],
        "summary": "Create a new post",
        "parameters": [
          {
            "name": "post",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Post"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Post created successfully",
            "schema": {
              "$ref": "#/definitions/Post"
            }
          }
        }
      },
      "put": {
        "tags": ["posts"],
        "summary": "Update an existing post",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "The ID of the post to update"
          },
          {
            "name": "post",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Post"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Post updated successfully",
            "schema": {
              "$ref": "#/definitions/Post"
            }
          }
        }
      },
      "delete": {
        "tags": ["posts"],
        "summary": "Delete a post",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "The ID of the post to delete"
          }
        ],
        "responses": {
          "200": {
            "description": "Post deleted successfully"
          }
        }
      }
    },
    "/products": {
      "get": {
        "tags": ["products"],
        "summary": "Get all products",
        "responses": {
          "200": {
            "description": "A list of products",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Product"
              }
            }
          }
        }
      },
      "post": {
        "tags": ["products"],
        "summary": "Create a new product",
        "parameters": [
          {
            "name": "product",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Product"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Product created successfully",
            "schema": {
              "$ref": "#/definitions/Product"
            }
          }
        }
      },
      "put": {
        "tags": ["products"],
        "summary": "Update an existing product",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "The ID of the product to update"
          },
          {
            "name": "product",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Product"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Product updated successfully",
            "schema": {
              "$ref": "#/definitions/Product"
            }
          }
        }
      },
      "delete": {
        "tags": ["products"],
        "summary": "Delete a product",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "The ID of the product to delete"
          }
        ],
        "responses": {
          "200": {
            "description": "Product deleted successfully"
          }
        }
      }
    },
    "/users": {
      "get": {
        "tags": ["users"],
        "summary": "Get all users",
        "responses": {
          "200": {
            "description": "A list of users",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/User"
              }
            }
          }
        }
      },
      "post": {
        "tags": ["users"],
        "summary": "Create a new user",
        "parameters": [
          {
            "name": "user",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "User created successfully",
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        }
      },
      "put": {
        "tags": ["users"],
        "summary": "Update an existing user",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "The ID of the user to update"
          },
          {
            "name": "user",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User updated successfully",
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        }
      },
      "delete": {
        "tags": ["users"],
        "summary": "Delete a user",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "The ID of the user to delete"
          }
        ],
        "responses": {
          "200": {
            "description": "User deleted successfully"
          }
        }
      }
    }
  },
  "definitions": {
    "Order": {
      "type": "object",
      "properties": {
        "userId": {
          "type": "string",
          "description": "The ID of the user who placed the order",
          "default": "677962da9577e3b04519b2fc"
        },
        "items": {
          "type": "array",
          "items": {
            "type": "object",
            "properties": {
              "productId": {
                "type": "string",
                "description": "The ID of the product",
                "default": "677962da9577e3b04519b2fc"
              },
              "quantity": {
                "type": "integer",
                "description": "The quantity of the product"
              }
            }
          }
        },
        "status": {
          "type": "string",
          "enum": ["pending", "completed", "cancelled"],
          "default": "pending"
        },
        "order_date": {
          "type": "string",
          "format": "date-time"
        }
      }
    },
    "Post": {
      "type": "object",
      "properties": {
        "title": {
          "type": "string",
          "description": "The title of the post"
        },
        "content": {
          "type": "string",
          "description": "The content of the post"
        },
        "mainImage": {
          "type": "string",
          "description": "The main image URL of the post"
        },
        "creatorId": {
          "type": "string",
          "description": "The ID of the creator",
          "default": "677962da9577e3b04519b2fc"
        },
        "images": {
          "type": "array",
          "items": {
            "type": "string"
          },
          "description": "A list of image URLs"
        },
        "status": {
          "type": "string",
          "enum": ["public", "private", "archive"],
          "default": "public"
        },
        "created_date": {
          "type": "string",
          "format": "date-time"
        }
      }
    },
    "Product": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "description": "The name of the product"
        },
        "price": {
          "type": "number",
          "description": "The price of the product"
        },
        "stock": {
          "type": "integer",
          "description": "The stock of the product"
        },
        "category": {
          "type": "string",
          "description": "The category of the product"
        },
        "tags": {
          "type": "array",
          "items": {
            "type": "string"
          },
          "description": "A list of tags associated with the product"
        }
      }
    },
    "User": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "description": "The name of the user"
        },
        "email": {
          "type": "string",
          "description": "The email of the user"
        },
        "phone": {
          "type": "string",
          "description": "The phone number of the user"
        },
        "address": {
          "type": "string",
          "description": "The address of the user"
        },
        "status": {
          "type": "string",
          "enum": ["deleted", "active", "blocked"],
          "default": "active"
        },
        "loyalityLevel": {
          "type": "integer",
          "description": "The loyalty level of the user"
        },
        "role": {
          "type": "string",
          "enum": ["user", "moderator", "administrator", "owner"],
          "default": "user"
        },
        "created_date": {
          "type": "string",
          "format": "date-time"
        }
      }
    }
  }
}